Data_Augmentation/fewshot_prompting/requirements.txt
langchain==1.0.0
langchain_community==0.4
langchain_ollama==1.0.0
pandas==2.3.3
flask==3.1.2
prompt_toolkit==3.0.52
langchain-classic==1.0.0
langchain-core==1.0.0
pyyaml==6.0.1
dspy-ai==3.0.4
